backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Google Gemini (Primary LLM + Embeddings)
google-generativeai==0.8.3

# OpenAI (Fallback LLM)
openai==1.54.3

# PDF / DOCX parsing
pdfplumber==0.11.4
python-docx==1.1.2

# Data / ML
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3

# HTTP / Scraping
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Async
anyio==4.6.2

# Utilities
python-jose==3.3.0
aiofiles==24.1.0
uuid==1.30
